requirements.txt
asgiref==3.4.1
attrs==21.2.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
cryptography==2.8
distro==1.4.0
Django==3.2.8
django-ckeditor==6.1.0
django-cleanup==5.2.0
django-core==1.4.1
django-cors-headers==3.10.0
django-environ==0.7.0
django-filter==21.1
django-js-asset==1.2.2
django-taggit==1.5.1
django-taggit-serializer==0.1.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
entrypoints==0.3
flake8==3.9.2
flake8-bugbear==21.9.2
flake8-builtins==1.5.3
flake8-django==1.1.2
flake8-functions-names==0.0.6
flake8-import-order==0.18.1
flake8-isort==4.0.0
flake8-polyfill==1.0.2
flake8-variables-names==0.0.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
gunicorn==20.1.0
httplib2==0.14.0
idna==2.8
lazr.uri==1.0.3
Markdown==3.3.4
mccabe==0.6.1
oauthlib==3.1.0
pbr==5.6.0
pep8-naming==0.12.1
Pillow==8.3.2
ptyprocess==0.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.2.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-environ==0.4.54
pytz==2019.3
PyYAML==5.3.1
requests==2.22.0
rsa==4.7.2
SecretStorage==2.3.1
six==1.14.0
sqlparse==0.2.4
testfixtures==6.18.3
testresources==2.0.1
uritemplate==3.0.1
urllib3==1.25.8
wadllib==1.3.3
websocket-client==0.57.0
whitenoise==5.2.0
